chore: explain env vars
