+++ src/constants.ts
@@ -1,7 +1,10 @@
+/** Used on owner-amoy.test.ts */
 export const RPC_AMOY_DEFAULT = process.env.RPC_AMOY || "https://polygon-amoy.drpc.org"
 export const CONTRACT_REGISTRY_AMOY_DEFAULT = process.env.CONTRACT_REGISTRY_AMOY || "0xBC56d0883ef228b2B16420E9002Ece0A46c893F8"
+/** Used on create.test.ts */
 export const AMOY_CHAIN_ID = 80002;
 
 export const RPC_CARDONA_DEFAULT = process.env.RPC_CARDONA || "https://endpoints.omniatech.io/v1/polygon-zkevm/testnet/public"
 export const CONTRACT_REGISTRY_CARDONA_DEFAULT = process.env.CONTRACT_REGISTRY_CARDONA || "0x03d5003bf0e79c5f5223588f347eba39afbc3818"
+/** Used on create.test.ts */
 export const CARDONA_CHAIN_ID = 2442;
